travel_planner/requirements.txt
# Core Dependencies
streamlit>=1.32.0
loguru>=0.7.2
pydantic>=2.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
aiocache>=0.12.2
pandas>=2.2.2
fpdf2>=2.7.8

# API Integrations
google-generativeai>=0.3.0
python-amadeus>=8.2.0
requests-oauthlib>=1.3.1
python-dateutil>=2.8.2

# Development & Testing
pytest>=8.3.2
pytest-asyncio>=0.23.7
pytest-cov>=4.1.0
black>=24.4.2
isort>=5.13.2
mypy>=1.10.0

# Async Utilities
aiohttp>=3.9.0
asyncio>=3.4.3

# Data Processing
numpy>=1.26.0
pandas>=2.2.2
